Added app manifest to make sure we always have admin access (for the httplistener)
Edited create account command and added in a reply for when the account has been created.
Changed all projects to build in 32bit mode (for edit & continue + nhibernate)
Do not ever run the base code for OnCommandFail and now output the error information to the correct channel as specified in our config instead.